Refactored the database queries into its own helper functions Offloaded server calculations into the db instead
