+++ server/app.js
@@ -1,30 +1,15 @@
-var mysql = require('mysql');
 var express = require('express');
-// var db = require('./database/config.js');
-var app = express();
 var bodyParser = require('body-parser');
-var env = process.env.NODE_ENV || 'development';
-var db = require('./database/config')[env];
+var app = express();
+var db = require('./database/helper.js');
 
 app.use(express.static(__dirname + '/../client/dist'));
 app.use(bodyParser.urlencoded({extended: true}));
 
-var connection = mysql.createConnection(db);
-
-connection.connect(function(err) {
-  if (err) {
-     console.log(err);
-  }
-  // console.log("Connected to db");
-});
-
 app.get('/reviews/audience/:title', (req, res) => {
-  console.log("Connected to db");
-  connection.query(`SELECT id, users.username, review, stars FROM audience_reviews \
-  INNER JOIN users ON users.user_id = audience_reviews.user_id 
-  AND movie_id=${req.params.title} LIMIT 4;`, (err, results) => {
+  db.getAudienceReview(req.params.title, (err, results) => {
     if (err) {
-      console.log("Error getting reviews", err);
+      throw err;
     } else {
       res.send(JSON.stringify(results, null, 2));
     }
@@ -32,29 +17,14 @@ app.get('/reviews/audience/:title', (req, res) => {
 });
 
 app.get('/reviews/scoreboard/:title', (req, res) => {
-  connection.query(`SELECT * FROM audience_reviews where movie_id=${req.params.title};`, (err, results) => {
+  db.getAudienceScoreboard(req.params.title, (err, results) => {
     if (err) {
-      console.log('Error getting reviews', err);
+      throw err;
     } else {
-      var reviewCount = 0;
-      var rating = 0;
-      var liked = 0;
-      results.forEach( (movie) => {
-        reviewCount += 1;
-        rating += movie.stars;
-        liked += movie.liked;
-      });
-      console.log('Number of reviews', reviewCount);
-      console.log('stars', rating/reviewCount);
-      console.log('% liked: ', liked/reviewCount);
-
-      var scoreboard = {};
-      scoreboard.audienceScore = (liked/reviewCount*100).toFixed(0);
-      scoreboard.averageRating = (rating/reviewCount).toFixed(1);
-      scoreboard.userRatings = reviewCount;
-      res.send(scoreboard);
+      res.send(JSON.stringify(results, null, 2));
     }
-  })
+  });
+  //scoreboard.audienceScore = (liked/reviewCount*100).toFixed(0);
 });
 
 let port = process.env.PORT || 9003;
@@ -63,4 +33,4 @@ var server = app.listen(port, function() {
   console.log(`listening on port ${port}`);
 });
 
-module.exports = {server, connection};
+module.exports = {server};
